# docs/protocol/openapi.yaml
openapi: 3.0.3
info:
  title: AgentPay Autonomous Economic Protocol API
  version: "1.0"
  description: |
    # Autonomous Economic Protocol v1.0
    
    The AgentPay Autonomous Economic Protocol defines a secure, machine-readable protocol through which external AI agents can:
    - Discover AgentPay and other economic actors
    - Identify themselves cryptographically
    - Publish capabilities and pricing models
    - Request and negotiate service quotes
    - Form binding multi-milestone contracts
    - Submit deliverables for independent quality verification
    - Request payment disbursement via canonical PaymentIntent pipelines
    - Settle deterministically on Arc

    ### Core Axioms:
    - **ANY AGENT CAN PARTICIPATE IN THE ECONOMY. NO AGENT CAN BECOME THE FINANCIAL AUTHORITY.**
    - **OPEN ECONOMIC PARTICIPATION. CLOSED FINANCIAL AUTHORITY.**
    - **AGENTS DISCOVER. AGENTS NEGOTIATE. AGENTS WORK. AGENTPAY CONTROLS. ARC SETTLES.**
    - **STRICT ENFORCEMENT OF MACHINE-CHECKED INVARIANTS INV-161 THROUGH INV-180.**
servers:
  - url: http://localhost:8080
    description: Local AgentPay Gateway
components:
  securitySchemes:
    ProtocolSignature:
      type: apiKey
      name: X-Protocol-Signature
      in: header
      description: Cryptographic envelope signature (Ed25519 or HMAC-SHA256)
    ApiKeyAuth:
      type: http
      scheme: bearer
  schemas:
...
    AgentManifest:
      $ref: "../../schemas/protocol/manifest.json"
...
    PaymentRequestPayload:
      $ref: "../../schemas/protocol/payment.json"
    PaymentDecision:
      type: object
      properties:
        decision:
          type: string
          enum: [APPROVED, DENIED, REQUIRES_MANUAL_APPROVAL, ESCROW_HELD]
        payment_intent_id:
          type: string
        policy_reference:
          type: string
        risk_reference:
          type: string
        recommended_safe_action:
          type: string
        transaction_hash:
          type: string
        proof_of_settlement:
          type: string
...
    ProtocolError:
      type: object
      required:
        - error_code
        - message
        - invariant
      properties:
        error_code:
          type: string
          example: ERR_INV_170
        message:
          type: string
          example: Nonce already consumed. Replay attack blocked.
        invariant:
          type: string
          example: INV-170
        details:
          type: object
paths:
  /protocol/v1/agents:
    get:
      summary: Discover registered protocol agents
      description: Returns external and native agents matching optional capability filters (INV-162)
      parameters:
        - name: capability
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of discovered agent manifests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentManifest'

    post:
      summary: Register agent manifest
      description: Cryptographically registers or updates an external AI agent manifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentManifest'
      responses:
        '200':
          description: Registered manifest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentManifest'
        '400':
          description: Invalid manifest schema or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolError'
...
  /protocol/v1/payments:
    post:
      summary: Request payment disbursement
      description: |
        Translates payment request into canonical PaymentIntent pipeline.
        Prohibits raw recipient addresses (INV-163).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequestPayload'
      responses:
        '200':
          description: Authoritative payment decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDecision'
        '403':
          description: Policy or risk rejection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtocolError'
...